chapter-generator/requirements.txt
# S3 호환 캐시 시스템을 위한 새로운 의존성
boto3>=1.26.0
botocore>=1.29.0

# 기존 의존성 (추가할 수 있는 항목들)
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
opencv-python>=4.5.0
mediapipe>=0.8.0
tensorflow>=2.8.0
tqdm>=4.62.0
scipy>=1.7.0
python-dotenv>=0.19.0 